Major backup update -- too big of a commit.
Refactored to have a DatabaseHelper wrapping the live database, so that it can load the live one, or other version of the database for backup/restore processing.

The singleton for the live database is now in the Helper.
Most classes should call the helper, and not the database as a result.

Expanding the cloud restore function to have a progress in the UI.
Support the update of the system portion of the database from assets, (or more but only asserts supported at the time).